+++ spammpack-C/spammpack/src/spamm_fortran_interface.c
@@ -174,7 +174,7 @@ spamm_chunk_get_matrix_dilated_interface (float **A, spamm_chunk_t **chunk)
  * @param chunk The chunk.
  */
 void
-spamm_chunk_get_norm_interface (float **norm, spamm_chunk_t **chunk)
+spamm_chunk_get_norm_interface (double **norm, spamm_chunk_t **chunk)
 {
   *norm = spamm_chunk_get_norm(*chunk);
 }
@@ -185,7 +185,7 @@ spamm_chunk_get_norm_interface (float **norm, spamm_chunk_t **chunk)
  * @param chunk The chunk.
  */
 void
-spamm_chunk_get_norm2_interface (float **norm2, spamm_chunk_t **chunk)
+spamm_chunk_get_norm2_interface (double **norm2, spamm_chunk_t **chunk)
 {
   *norm2 = spamm_chunk_get_norm2(*chunk);
 }
